Change the data type of the upper shared memory limits from a signed integer to an unsigned long. This lifts variables like the maximum number of pages available for shared memory from 2^31 to 2^32 on 32 bit architectures, and from 2^31 to 2^64 on 64 bit architectures.
It should be noted that this changes breaks ABI on 64 bit architectures
because the size of the shmmax, shmmin, shmmni, shmseg and shmall members
of the shminfo structure has changed.

Silence on:	current@


Former-commit-id: fc4e53728e09160c6589a7d4c31136d3ffd59eba

+++ usr.bin/ipcs/ipcs.c
@@ -383,7 +383,7 @@ define|#
 directive|define
 name|SHMINFO_XVEC
 define|\
-value|X(shmmax, sizeof(int))				\ X(shmmin, sizeof(int))				\ X(shmmni, sizeof(int))				\ X(shmseg, sizeof(int))				\ X(shmall, sizeof(int))
+value|X(shmmax, sizeof(u_long))				\ X(shmmin, sizeof(u_long))				\ X(shmmni, sizeof(u_long))				\ X(shmseg, sizeof(u_long))				\ X(shmall, sizeof(u_long))
 end_define
 
 begin_define
